Revert to vanilla Mantine v7 theme
Removes the extended design language system introduced in commit 5cde6bc, reverting to a simpler vanilla Mantine v7 setup. The complex theme customizations were deemed unnecessary for this project.

Changes:
- Remove custom theme files (clientTheme, themeAlert, themeButton, tailwindTheme)
- Remove SCSS partials and globals.scss
- Remove BodyText component and ClientThemeProvider
- Restore simple MantineThemeOverride in lib/theme.ts
- Update app/layout.tsx to use basic MantineProvider
- Revert tailwind.config.ts to simple configuration
- Create styles/globals.css with basic Tailwind imports
- Remove sass, @emotion/react, @emotion/styled dependencies
- Update all imports from @/theme to @/lib
