+++ genai-cookbook/app/cookbook/[recipe]/layout.tsx
@@ -2,7 +2,7 @@
 
 import { Flex } from '@mantine/core'
 
-import { appShellContentHeight } from '@/theme/theme'
+import { appShellContentHeight } from '@/lib/theme'
 import { Toolbar } from '@/components/Toolbar'
 import { redirect } from 'next/navigation'
 import { cookbookRoute } from '@/lib/constants'
