+++ genai-cookbook/styles/globals.css
@@ -1,5 +1,3 @@
-@use '@/theme/partials/_base' as *;
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -9,5 +7,3 @@
         text-wrap: balance;
     }
 }
-
-@include base;
