--- genai-cookbook/theme/partials/_font.scss
@@ -1,159 +1,0 @@
-@use '_breakpoints' as *;
-@use '_mantine' as *;
-
-@tailwind utilities;
-
-$roboto: var(--roboto);
-$inter: var(--inter);
-
-// Custom font style classnames
-@mixin h-60 {
-    font-size: rem(60);
-    font-weight: 500;
-    line-height: rem(80);
-    letter-spacing: rem(-4);
-}
-
-@mixin h-48 {
-    font-size: rem(48);
-    font-weight: 500;
-    line-height: rem(56);
-    letter-spacing: rem(-1.92);
-}
-
-@mixin h-40 {
-    font-size: rem(40);
-    font-weight: 500;
-    line-height: rem(48);
-    letter-spacing: rem(-2);
-}
-
-@mixin h-32 {
-    font-size: rem(32);
-    font-weight: 500;
-    line-height: rem(40);
-    letter-spacing: rem(-2);
-}
-
-@mixin h-30 {
-    font-size: rem(30);
-    font-weight: 500;
-    line-height: rem(40);
-    letter-spacing: rem(-1.5);
-}
-
-@mixin h-28 {
-    font-size: rem(28);
-    font-weight: 500;
-    line-height: rem(36);
-    letter-spacing: rem(-1.5);
-}
-
-@mixin body-24 {
-    font-size: rem(24);
-    font-weight: 500;
-    line-height: rem(32);
-    letter-spacing: rem(-0.6);
-}
-
-@mixin body-20 {
-    font-size: rem(20);
-    font-weight: 500;
-    line-height: rem(26);
-    letter-spacing: rem(-0.5);
-}
-
-@mixin body-18 {
-    font-size: rem(18);
-    font-weight: 500;
-    line-height: rem(24);
-    letter-spacing: rem(-0.72);
-}
-
-@mixin body-16 {
-    font-size: rem(16);
-    font-weight: 500;
-    line-height: rem(24);
-    letter-spacing: rem(-0.32);
-}
-
-@mixin body-14 {
-    font-size: rem(14);
-    font-weight: 500;
-    line-height: rem(20);
-    letter-spacing: rem(-0.35);
-}
-
-@mixin body-12 {
-    font-size: rem(12);
-    font-weight: 500;
-    line-height: rem(20);
-    letter-spacing: rem(-0.3);
-}
-
-@mixin applyFonts {
-    .h-60 {
-        @include h-60;
-    }
-    .h-48 {
-        @include h-48;
-    }
-    .h-40 {
-        @include h-40;
-    }
-    .h-32 {
-        @include h-32;
-    }
-    .h-30 {
-        @include h-30;
-    }
-    .h-28 {
-        @include h-28;
-    }
-    .body-24 {
-        @include body-24;
-    }
-    .body-24-light {
-        @include body-24;
-        font-weight: 400;
-    }
-    .body-20 {
-        @include body-20;
-    }
-    .body-20-light {
-        @include body-20;
-        font-weight: 400;
-    }
-    .body-18 {
-        @include body-18;
-    }
-    .body-18-light {
-        @include body-18;
-        font-weight: 400;
-    }
-    .body-16 {
-        @include body-16;
-    }
-    .body-16-light {
-        @include body-16;
-        font-weight: 400;
-    }
-    .body-14 {
-        @include body-14;
-    }
-    .body-14-light {
-        @include body-14;
-        font-weight: 400;
-    }
-    .body-12 {
-        @include body-12;
-    }
-    .body-12-light {
-        @include body-12;
-        font-weight: 400;
-    }
-}
-
-@layer utilities {
-    @include applyFonts;
-}
