--- genai-cookbook/theme/partials/_mantine.scss
@@ -1,64 +1,0 @@
-@use 'sass:math';
-
-// Define variables for your breakpoints,
-// values must be the same as in your theme
-$mantine-breakpoint-xs: '36em';
-$mantine-breakpoint-sm: '48em';
-$mantine-breakpoint-md: '62em';
-$mantine-breakpoint-lg: '75em';
-$mantine-breakpoint-xl: '88em';
-
-@function rem($value) {
-    @return #{math.div(math.div($value, $value * 0 + 1), 16)}rem;
-}
-
-@mixin light {
-    [data-mantine-color-scheme='light'] & {
-        @content;
-    }
-}
-
-@mixin dark {
-    [data-mantine-color-scheme='dark'] & {
-        @content;
-    }
-}
-
-@mixin hover {
-    @media (hover: hover) {
-        &:hover {
-            @content;
-        }
-    }
-
-    @media (hover: none) {
-        &:active {
-            @content;
-        }
-    }
-}
-
-@mixin smaller-than($breakpoint) {
-    @media (max-width: $breakpoint) {
-        @content;
-    }
-}
-
-@mixin larger-than($breakpoint) {
-    @media (min-width: $breakpoint) {
-        @content;
-    }
-}
-
-// Add direction mixins if you need rtl support
-@mixin rtl {
-    [dir='rtl'] & {
-        @content;
-    }
-}
-
-@mixin ltr {
-    [dir='ltr'] & {
-        @content;
-    }
-}
