--- genai-cookbook/theme/tailwindTheme.js
@@ -1,78 +1,0 @@
-module.exports = {
-    extend: {
-      colors: {
-        black: 'var(--Black)',
-        white: 'var(--White)',
-        red: 'var(--Red)',
-        'color-black': 'var(--Color-Black)',
-        'color-white': 'var(--Color-White)',
-        'Elements-Green-100-140': 'var(--Elements-Green-100-140)',
-        'Elements-Green-120': 'var(--Elements-Green-120)',
-        'Elements-Green-140-100': 'var(--Elements-Green-140-100)',
-        'Elements-Green-20-60': 'var(--Elements-Green-20-60)',
-        'Elements-Green-20-Alpha': 'var(--Elements-Green-20-Alpha)',
-        'Elements-Neb-100': 'var(--Elements-Neb-100)',
-        'Elements-Neb-20-Alpha': 'var(--Elements-Neb-20-Alpha)',
-        'Elements-Neb-30-Alpha': 'var(--Elements-Neb-30-Alpha)',
-        'Elements-Neb-40-Alpha': 'var(--Elements-Neb-40-Alpha)',
-        'Elements-Neb-60-Alpha': 'var(--Elements-Neb-60-Alpha)',
-        'Elements-Neb-Ultra': 'var(--Elements-Neb-Ultra)',
-        'Elements-Neb-Ultra-Super': 'var(--Elements-Neb-Ultra-Super)',
-        'Elements-Orange-100': 'var(--Elements-Orange-100)',
-        'Elements-Red-110': 'var(--Elements-Red-110)',
-        'Elements-Red-20-60': 'var(--Elements-Red-20-60)',
-        'Elements-Twilight-0-100': 'var(--Elements-Twilight-0-100)',
-        'Elements-Twilight-0-70': 'var(--Elements-Twilight-0-70)',
-        'Elements-Twilight-0-80': 'var(--Elements-Twilight-0-80)',
-        'Elements-Twilight-0-90': 'var(--Elements-Twilight-0-90)',
-        'Elements-Twilight-10-90': 'var(--Elements-Twilight-10-90)',
-        'Elements-Twilight-100-40': 'var(--Elements-Twilight-100-40)',
-        'Elements-Twilight-20': 'var(--Elements-Twilight-20)',
-        'Elements-Twilight-20-70': 'var(--Elements-Twilight-20-70)',
-        'Elements-Twilight-30': 'var(--Elements-Twilight-30)',
-        'Elements-Twilight-30-60': 'var(--Elements-Twilight-30-60)',
-        'Elements-Twilight-30-70': 'var(--Elements-Twilight-30-70)',
-        'Elements-Twilight-30-80': 'var(--Elements-Twilight-30-80)',
-        'Elements-Twilight-40-60': 'var(--Elements-Twilight-40-60)',
-        'Elements-Twilight-5-100': 'var(--Elements-Twilight-5-100)',
-        'Elements-Twilight-5-80': 'var(--Elements-Twilight-5-80)',
-        'Elements-Twilight-5-90': 'var(--Elements-Twilight-5-90)',
-        'Elements-Twilight-50': 'var(--Elements-Twilight-50)',
-        'Elements-Twilight-Alpha': 'var(--Elements-Twilight-Alpha)',
-        'Elements-Twilight-60': 'var(--Elements-Twilight-60)',
-        'Elements-Twilight-60-40': 'var(--Elements-Twilight-60-40)',
-        'Elements-Twilight-60-50': 'var(--Elements-Twilight-60-50)',
-        'Elements-Twilight-70-20': 'var(--Elements-Twilight-70-20)',
-        'Elements-Twilight-70-50': 'var(--Elements-Twilight-70-50)',
-        'Elements-Twilight-90-60': 'var(--Elements-Twilight-90-60)',
-      },
-      fontFamily: {
-        roboto: 'var(--roboto)',
-        inter: 'var(--inter)',
-      },
-      boxShadow: {
-        1: 'var(--Shadow-1)',
-        2: 'var(--Shadow-2)',
-        3: 'var(--Shadow-3)',
-        4: 'var(--Shadow-4)',
-        5: 'var(--Shadow-5)',
-        6: 'var(--Shadow-6)',
-      },
-      borderWidth: {
-        3: '3px',
-      },
-      spacing: {
-        60: '240px',
-        tablet: '768px',
-        desktop: '1280px',
-      },
-    },
-    screens: {
-      xs: '576px',
-      sm: '640px',
-      md: '768px',
-      lg: '1024px',
-      xl: '1280px',
-    },
-  };
-
